pyproject.toml
[project]
name = "ytchatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.1",
    "uvicorn>=0.38.0",
    "youtube-transcript-api>=1.2.3",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "python-dotenv>=1.0.0",
    "langgraph>=1.0.3",
    "langchain-core>=1.0.4",
]
